EDIT VideoDetail, userDetail

+++ controllers/videoControllers.js
@@ -69,8 +69,11 @@ export const getEditVideo = async (req, res) => {
     } = req;
     try {
         const video = await Video.findById(id);
-        console.log(video);
-        res.render("editVideo", {pageTitle: `Edit ${video.title}`, video});
+        if ( video.creator !== req.user.id ) {
+            throw Error();
+        } else {
+            res.render("editVideo", {pageTitle: `Edit ${video.title}`, video});
+        }
     } catch (error) {
         res.redirect(routes.home);
     }
@@ -96,7 +99,11 @@ export const deleteVideo = async (req, res) => {
     } = req;
 
     try {
-        await Video.findOneAndRemove({_id: id});
+        if ( video.creator !== req.user.id ) {
+            throw Error();
+        } else {
+            await Video.findOneAndRemove({_id: id});
+        }
     } catch (error) {
         console.log(error);
     }
